Update ScriptProcessMonitor activity to Android Lollipop
Firstly, the Toolbar was added since this activity had lost its ActionBar so the activity showed only a ListView. Secondly, the aforementioned ListView was replaced by a list implemented using RecyclerView. Thirdly, ScriptMonitorListAdapter was created to serve as the adapter between the RecyclerView and its corresponding data (the list of processes running currently).